Changed structure and made an interface for packet callbacks
