+++ Packets/ClientAccepted.cs
@@ -12,11 +12,11 @@ namespace SMServer.Packets
         {
         }
 
-        public virtual void Serialize(BigEndianBinaryWriter writer)
+        public void Serialize(BigEndianBinaryWriter writer)
         {
         }
 
-        public virtual void Deserialize(BigEndianBinaryReader reader)
+        public void Deserialize(BigEndianBinaryReader reader)
         {
         }
     }
