+++ Packets/JoinConfirmation.cs
@@ -4,17 +4,17 @@
     internal class JoinConfirmation : IPacket
     {
         public const byte PacketId = 10;
-
+        
         // Constructor
         public JoinConfirmation()
         {
         }
 
-        public virtual void Serialize(BigEndianBinaryWriter writer)
+        public void Serialize(BigEndianBinaryWriter writer)
         {
         }
 
-        public virtual void Deserialize(BigEndianBinaryReader reader)
+        public void Deserialize(BigEndianBinaryReader reader)
         {
         }
     }
